fix paths and rename controller file to videosController
