+++ api/routes/movies.js
@@ -1,7 +1,7 @@
 // create an isolated instance of a router
 // allows to create routes that are mountable to a specific url path
 const router = require("express").Router();
-const userController = require("../controllers/moviesController");
+const userController = require("../controllers/videosController");
 
 // middleware that is specific to this router
 const timeLog = (req, res, next) => {
@@ -12,7 +12,7 @@ router.use(timeLog);
 
 // home route will call the getTrendingVideos function
 // when the route is hit
-router.route("/").get(userController.getTrendingVideos);
+router.route("/trending-videos").get(userController.getTrendingVideos);
 
 // use node module to export the route
 module.exports = router;
