+++ api/server.js
@@ -7,7 +7,7 @@ app.use(cors());
 // get the route handlers
 const movies = require("./routes/movies");
 // mount the movies functions to the /movies route
-app.use("/movies", movies);
+app.use("/videos", movies);
 
 app.get("/", (req, res) => {
   res.status(200).json("Welcome to the API"); // Sets status to 201 (Created)
